requirements.txt
langgraph>=0.0.15
langchain>=0.0.335
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pydantic>=2.5.0
Pillow>=9.0.0  # For image processing
openai>=1.0.0